\name{siren-internal}
\alias{siren-internal}
\alias{acquiextraction}
\alias{congru}
\alias{extensionacq}
\alias{extensionacqm}
\alias{ordnormulti}
\alias{pconmul}
\alias{ppnd}
\alias{selecore}
\alias{transpose}
\alias{size}
\title{
siren-internal
}
\description{
These functions are for internal use only.
}
\usage{
acquiextraction
congru
extensionacq
extensionacqm
ordnormulti
pconmul
ppnd
selecore
transpose
size
}
\author{
David Navarro-Gonzalez

Pere J. Ferrando

Fabia Morales-Vives

Ana Hernandez-Dorado
}
\keyword{ internal }
