\name{siren-package}
\alias{siren-package}
\title{
An hybrid CFA-EFA R package for controlling acquiescence in restricted solutions
}
\description{
Hybrid multi-stage factor analytic procedure for controlling the effects of acquiescence in multidimensional questionnaires for which a restricted factorial solution (Ferrando & Lorenzo-Seva, 2000) can be specified.
}
\value{
\item{\code{\link{acquihybrid}}}{Performs hybrid multi-stage factor analytic procedure for controlling acquiescence in restricted solutions.}
}
\references{
Ferrando, P. J., & Seva, U. L. (2000). Unrestricted versus restricted factor analysis of multidimensional test items: Some aspects of the problem and some suggestions. Psicologica, 21(2), 301-323.
}
\author{
David Navarro-Gonzalez

Pere J. Ferrando

Fabia Morales-Vives

Ana Hernandez-Dorado
}
\keyword{ package }
