\name{siren-package}
\alias{siren-package}
\title{
An hybrid CFA-EFA R package for controlling acquiescence in restricted solutions
}
\description{
Hybrid multi-stage factor analytic procedure for controlling the effects of acquiescence in multidimensional questionnaires for which a restricted factorial solution (Ferrando & Lorenzo-Seva, 2000) can be specified.
}
\value{
\item{\code{\link{acquihybrid}}}{Performs hybrid multi-stage factor analytic procedure for controlling acquiescence in restricted solutions.}
}
\references{
Ferrando, P. J., & Seva, U. L. (2000). Unrestricted versus restricted factor analysis of multidimensional test items: Some aspects of the problem and some suggestions. Psicologica, 21(2), 301-323.
Ferrando, P. J., & Lorenzo-Seva, U. (2010). Unrestricted item factor analysis and some relations with item response theory. Technical report, Department of Psychology, Universitat Rovira i Virgili, Tarragona.http://psico.fcep.urv.es/utilitats/factor.
Ferrando, P. J., Lorenzo-Seva, U., & Chico, E. (2003). Unrestricted factor analytic procedures for assessing acquiescent responding in balanced, theoretically unidimensional personality scales. Multivariate Behavioral Research, 3(38):353–374, doi:10.1207/S15327906MBR3803_04
}
\author{
David Navarro-Gonzalez

Pere J. Ferrando

Fabia Morales-Vives

Ana Hernandez-Dorado
}
\keyword{ package }
